refactor: restructured code
